Fixed Mongo serizlization
