+++ brvs-core/src/main/java/iroha/validation/transactions/storage/impl/dummy/DummyMemoryTransactionVerdictStorage.java
@@ -1,4 +1,4 @@
-package iroha.validation.transactions.storage.impl;
+package iroha.validation.transactions.storage.impl.dummy;
 
 import io.reactivex.Observable;
 import io.reactivex.subjects.PublishSubject;
